Fix e2e test
